fixed bug to check size of array

+++ sort.h
@@ -19,6 +19,6 @@ typedef struct listint_s
 
 void print_array(const int *array, size_t size);
 void print_list(const listint_t *list);
-void swap(int *a, int *b);
+void swap(size_t *a, size_t *b);
 void bubble_sort(int *array, size_t size);
 #endif
